Stops counting clicks wins and new game button
Added gameStatus that will either be active or inactive. When a game is
over (win/tie) the gameStatus will change to inactive and will be
active once you click on Start New Game button. Start New Game button
will clear the game board and change the status of the game to active.