Displays number of complete games and wrote README
I added a button to access the get games function that will show the
completed games (over = true) for the user that is signed in. The README
file has also been updated to describe my plan and strategy, the
technologies I used in making the tic-tac-toe game, link to my
wireframe, and user stories.